+++ reusables/multiprocess_helpers.py
@@ -210,12 +210,25 @@ class Tasker(object):
         else:
             _logger.warning("Received an unknown command '{0}'".format(cmd))
 
+    def hook_pre_command(self):
+        pass
+
+    def hook_post_command(self):
+        pass
+
+    def hook_pre_task(self):
+        pass
+
+    def hook_post_task(self):
+        pass
+
     def main_loop(self, stop_at_empty=False):
         """Blocking function that can be run directly, if so would probably
         want to specify 'stop_at_empty' to true, or have a separate process
         adding items to the queue. """
         try:
             while True:
+                self.hook_pre_command()
                 self._check_command_queue()
                 if self.run_until and self.run_until < _datetime.datetime.now():
                     _logger.info("Time limit reached")
@@ -225,6 +238,7 @@ class Tasker(object):
                 if self._pause.value:
                     _time.sleep(.5)
                     continue
+                self.hook_post_command()
                 self._update_tasks()
                 task_id = self._free_task()
                 if task_id:
@@ -235,12 +249,15 @@ class Tasker(object):
                             break
                         self._return_task(task_id)
                     else:
+                        self.hook_pre_task()
                         _logger.debug("Starting task on {0}".format(task_id))
                         try:
                             self._start_task(task_id, task)
                         except Exception as err:
                             _logger.exception("Could not start task {0} -"
                                               " {1}".format(task_id, err))
+                        else:
+                            self.hook_post_task()
         finally:
             _logger.info("Ending main loop")